Streamline processing from bands
-modified frequency_bands_obj() method of decimation_level.py to
remove duplication of band_edges assignment
-tried to remove all calls to deprecated from_emtf_band_df() method

[Issue(s): #150]

+++ aurora/sandbox/io_helpers/emtf_band_setup.py
@@ -75,8 +75,8 @@ class EMTFBandSetupFile:
 
         """
         band_edges = {}
-        lower_edges = pd.Series(index=self.df.index)
-        upper_edges = pd.Series(index=self.df.index)
+        lower_edges = pd.Series(index=self.df.index, dtype="float64")
+        upper_edges = pd.Series(index=self.df.index, dtype="float64")
         if not self.sample_rate:
             print("cannot define frequencies of sample rate undefined")
             raise Exception
